refactor: reorganize async utilities and tests
- Move async-related type utilities and functions to a dedicated folder structure under `src/common/functions`.
- Implement `isAsyncConfig` and `isAsyncFnMap` runtime validation functions.
- Create `expandFnMap` function for handling function maps with async capabilities.
- Update imports across the codebase to reflect the new structure.
- Remove deprecated files and tests related to async utilities.
- Ensure type tests for `IsAsyncConfig` and `IsAsyncFnMap` are preserved in the new location.

_NO_CI
