+++ src/machine/functions/index.ts
@@ -1,5 +1,4 @@
 export * from './create';
-export * from './expandFnMap';
 export * from './subcriber';
 export * from './to';
 export * from './src';
